OAK-11607 get rid of the Iterable implementation, just use the iterator
